Refactoring, header and footer now part of page
