+++ Frontend/blog-next-frontend/pages/about.js
@@ -1,12 +1,8 @@
-import ResponsiveAppBar from "../resource/components/global/headerBar";
+import ResponsiveAppBar from "../resource/layout/headerBar";
 import { Box } from "@mui/system";
+import Page from "../resource/layout/page";
 
 export function About() {
-    return (
-        <Box>
-          <ResponsiveAppBar />
-        </Box>
-
-    );
+  return <Page></Page>;
 }
 export default About;
