--- Frontend/blog-next-frontend/pages/browse.js
@@ -1,12 +1,0 @@
-import ResponsiveAppBar from "../resource/components/global/headerBar";
-import { Box } from "@mui/system";
-
-export function Browse() {
-    return (
-        <Box>
-          <ResponsiveAppBar />
-        </Box>
-
-    );
-}
-export default Browse;
