+++ Frontend/blog-next-frontend/pages/index.js
@@ -1,16 +1,15 @@
 import styles from "../styles/Home.module.css";
 import { Box, Stack, Typography } from "@mui/material";
 
-import ResponsiveAppBar from "../resource/components/global/headerBar";
+import ResponsiveAppBar from "../resource/layout/headerBar";
 import GenericCard from "../resource/components/global/card";
 import React from "react";
-import Page from "../resource/components/page";
+import Page from "../resource/layout/page";
 
 export default function Home() {
   return (
     <>
       <Page title="Index">
-        <ResponsiveAppBar />
         <Box mx={"10%"} my="10%">
           <Typography variant="h1"> Welcome to the blog! </Typography>
           <Stack
