+++ Frontend/blog-next-frontend/pages/profile/settings.js
@@ -1,13 +1,12 @@
-import ResponsiveAppBar from "../../resource/components/global/headerBar";
+import ResponsiveAppBar from "../../resource/layout/headerBar";
 import { Box } from "@mui/system";
-import Page from "../../resource/components/page";
+import Page from "../../resource/layout/page";
 import Users from "../../data/mock_db/users";
 
 export function Settings() {
   let user = Users[0]; //TODO: Auth
   return (
     <Page title={"Settings: " + user.username}>
-      <ResponsiveAppBar />
       <Box></Box>
     </Page>
   );
